Refactor Person struct and adjust associated functions
In this commit, the `Person` structure has been refactored to accommodate an array of `Address` structs instead of an array of strings. The `Location` struct has been renamed to `Address`. Adjustments have been made to the `DisplayDetails()` function and the creation of a `Person` in `main()` to conform to these changes.