Improve details display
The Person struct has been refactored to use a slice of strings for addresses instead of an Address struct. The Address struct has been renamed to Location, which preserves its original fields. The method 'displayCity' has been added to the Person struct to isolate the logic for retrieving a city. The details display in the 'DisplayDetails' method is now formatted using a constant.